+++ uTransnet-Calc/Assets/uTrans/Scripts/Data/LinkDTO.cs
@@ -5,15 +5,15 @@ namespace uTrans.Data
     public class LinkDTO : BaseDTO
     {
         [PrimaryKey, AutoIncrement]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Indexed]
-        public int ProjectId  { get; set; }
+        public long ProjectId  { get; set; }
 
 
 
-        public int FirstPointId { get; set; }
+        public long FirstPointId { get; set; }
 
-        public int SecondPointId { get; set; }
+        public long SecondPointId { get; set; }
     }
 }
